feat(settlement): add admin mutations for settlement attempts and job results
Add the minimum settlement admin surface to the admin JSON-RPC API,
serving the mutation slice of the settlement admin design (the
remainder — pause/quiesce, observability reads, audit log — stays with
the design split from #1562):

- admin_insertSettlementAttempt: append one new attempt to a job
  (insert-only, store-assigned attempt number; sender/nonce/fees are
  resolved from L1 by tx hash when omitted), for porting
  externally-submitted settlement transactions.
- admin_markSettlementAttemptDefinitelyFailed: overwrite an attempt's
  result with an operator assertion (new ClientErrorType variant
  AbandonedByAdmin), bypassing the upgrade-only rule. Re-drive is
  emergent from the run loop; the job never fails.
- admin_removeSettlementAttemptResult: delete an attempt result so the
  reloaded task re-derives it from L1 (the undo of the above).
- admin_forceRemoveSettlementJobResult: un-complete a job (drop its
  stale watcher, respawn a fresh task that re-drives it).
- admin_abortSettlementTask / admin_reloadSettlementTask: expose the
  existing task controls.

The attempt mutations take a trailing "force=true" parameter to edit
jobs that already have a terminal result, so attempt corrections can
land before a job-result removal (the respawned task could otherwise
re-derive the old result from uncorrected attempts).

Mutations are declarative over stored state: the storage write happens
under the per-job lock, then the live task (if any) is told to reload;
the response reports notified/absent/notify-failed. The regular result
writer now keeps an admin-abandoned result over weaker client-side
notes (on-chain evidence still supersedes it) so a task that has not
observed an override yet cannot wedge on the conflict.

Tests for touched files move from inline mod tests to sibling tests.rs
files.

Refs #1254
