+++ crates/agglayer-jsonrpc-api/src/testutils.rs
@@ -1,7 +1,11 @@
 use std::{future::IntoFuture as _, net::SocketAddr, sync::Arc};
 
-use agglayer_config::Config;
+use agglayer_config::{
+    settlement_service::{SettlementServiceConfig, SettlementTransactionConfig},
+    Config,
+};
 use agglayer_contracts::L1RpcClient;
+use agglayer_settlement_service::SettlementService;
 use agglayer_storage::{
     backup::BackupClient,
     stores::{debug::DebugStore, epochs::EpochsStore, pending::PendingStore, state::StateStore},
@@ -9,11 +13,13 @@ use agglayer_storage::{
 };
 use agglayer_types::{Certificate, CertificateId, Height, NetworkId};
 use alloy::{
+    network::EthereumWallet,
     providers::{
         fillers::{BlobGasFiller, ChainIdFiller, FillProvider, GasFiller, JoinFill, NonceFiller},
         mock::Asserter,
         Identity, ProviderBuilder, RootProvider,
     },
+    signers::local::PrivateKeySigner,
     transports::mock::MockTransport,
 };
 use jsonrpsee::http_client::{HttpClient, HttpClientBuilder};
@@ -148,6 +154,29 @@ impl TestContext {
         // Create AgglayerImpl
         let agglayer_impl = crate::AgglayerImpl::new(v0_service, rpc_service);
 
+        // A settlement service over the (empty) state store, with its own
+        // wallet-carrying provider pointed at a dead endpoint: startup
+        // recovery finds no jobs, and admin methods that reach L1 error out
+        // instead of hanging.
+        let settlement_provider = ProviderBuilder::new()
+            .wallet(EthereumWallet::from(
+                PrivateKeySigner::from_slice(&[0x11; 32]).expect("valid test signing key"),
+            ))
+            .connect_http(
+                "http://127.0.0.1:0"
+                    .parse()
+                    .expect("test provider URL should parse"),
+            );
+        let settlement_service = SettlementService::start(
+            SettlementServiceConfig::default(),
+            Arc::new(SettlementTransactionConfig::default()),
+            Arc::new(settlement_provider),
+            state_store.clone(),
+            cancellation_token.clone(),
+        )
+        .await
+        .expect("settlement service should start");
+
         // Create the routers
         let router = agglayer_impl.start().await.unwrap();
         let admin_router = AdminAgglayerImpl::new(
@@ -156,6 +185,7 @@ impl TestContext {
             state_store.clone(),
             debug_store.clone(),
             config.clone(),
+            settlement_service,
         )
         .start()
         .await
