+++ crates/agglayer-storage/src/tests/mocks/state_store.rs
@@ -9,8 +9,9 @@ use crate::{
     columns::latest_settled_certificate_per_network::SettledCertificate,
     error::Error,
     stores::{
-        MetadataReader, MetadataWriter, NetworkInfoReader, SettlementReader, SettlementWriter,
-        StateReader, StateWriter, UpdateEvenIfAlreadyPresent, UpdateStatusToCandidate,
+        EditEvenIfCompleted, MetadataReader, MetadataWriter, NetworkInfoReader, SettlementReader,
+        SettlementWriter, StateReader, StateWriter, UpdateEvenIfAlreadyPresent,
+        UpdateStatusToCandidate,
     },
 };
 mock! {
@@ -184,6 +185,33 @@ mock! {
             settlement_job_id: &SettlementJobId,
             attempt_sequence_number: u64,
             tx_result: &SettlementAttemptResult,
+        ) -> Result<(), Error>;
+
+        fn admin_insert_settlement_attempt(
+            &self,
+            settlement_job_id: &SettlementJobId,
+            settlement_attempt: &SettlementAttempt,
+            edit_even_if_completed: EditEvenIfCompleted,
+        ) -> Result<u64, Error>;
+
+        fn admin_override_settlement_attempt_result(
+            &self,
+            settlement_job_id: &SettlementJobId,
+            attempt_number: u64,
+            tx_result: &SettlementAttemptResult,
+            edit_even_if_completed: EditEvenIfCompleted,
+        ) -> Result<(), Error>;
+
+        fn admin_remove_settlement_attempt_result(
+            &self,
+            settlement_job_id: &SettlementJobId,
+            attempt_number: u64,
+            edit_even_if_completed: EditEvenIfCompleted,
+        ) -> Result<(), Error>;
+
+        fn admin_force_remove_settlement_job_result(
+            &self,
+            settlement_job_id: &SettlementJobId,
         ) -> Result<(), Error>;
     }
 }
