+++ proto/agglayer/storage/v0/settlement.proto
@@ -69,6 +69,9 @@ enum ClientErrorType {
 
   // Settlement succeeded in another tx.
   CLIENT_ERROR_TYPE_SETTLEMENT_SUCCEEDED_ELSEWHERE = 2;
+
+  // An administrator asserted that this attempt will never land on L1.
+  CLIENT_ERROR_TYPE_ABANDONED_BY_ADMIN = 3;
 }
 
 // Result for a successfully-executed contract call.
